add pagination and create func to generate jwt

+++ app/core/api_key_and_rate_limiting.py
@@ -7,9 +7,9 @@ cur = conn.cursor()
 cur.execute(
 	"""
 	CREATE TABLE IF NOT EXISTS api_key (
-		id_ TEXT PRIMARY KEY, 
-		api_key TEXT, 
-		user_id INTEGER, 
+		id_ TEXT PRIMARY KEY,
+		api_key TEXT,
+		user_id INTEGER,
 		expires_at REAL,
 		max_limit INTEGER CHECK (max_limit > 0),
 		current_limit_count INTEGER DEFAULT 0,
@@ -26,27 +26,34 @@ cur.execute(
 conn.commit()
 
 def api_limit_manage(user_id):
-	current_time = time.time()
 	cur.execute(
 		"""SELECT max_limit, current_limit_count, expires_at 
 		FROM api_key WHERE user_id=?""", (user_id,)
 		)
 	row = cur.fetchone()
+
+	if not row:
+		raise HTTPException(
+				status_code=400,
+				detail="An error occured while verifying jwt."
+			)
+
 	if row:
+		current_time = time.time()
 		max_limit, current_limit_count, expires_at = row
 
 		if current_limit_count < max_limit and current_time < expires_at:
 			new_count = current_limit_count + 1
 			
 			cur.execute(
-				"""UPDATE api_key 
+				"""UPDATE api_key
 					SET current_limit_count = ?
 					WHERE user_id = ?""",
 				(new_count, user_id)
 			)
 			conn.commit()
 			return None
-		
+
 		else:
 			cur.execute(
 				"DELETE FROM api_key WHERE user_id=?", (user_id)
@@ -57,18 +64,11 @@ def api_limit_manage(user_id):
 				detail="Api key expired or limit reached."
 			)
 
-	if not row:
-		raise HTTPException(
-				status_code=400,
-				detail="An error occured while logging in."
-			)
-
-
 def api_key_set(id_, api_key, user_id, max_limit, ttl):
 	expires_at = time.time() + ttl
 	cur.execute(
-		"""INSERT OR REPLACE INTO api_key (
-		id_, api_key, user_id, expires_at, max_limit) VALUES (?, ?, ?, ?)""",
-		(id_, api_key, user_id, expires_at, max_limit)
+			"""INSERT OR REPLACE INTO api_key (
+			id_, api_key, user_id, expires_at, max_limit) VALUES (?, ?, ?, ?)""",
+			(id_, api_key, user_id, expires_at, max_limit)
 		)
 	conn.commit()
